Style as a pop art image

Style as a pop art image

style_pop(
  data,
  palette = "base",
  colour = "ind",
  alpha = 0.3,
  fade = 0,
  background = "warhol",
  adjust = 0.7,
  panels = 4,
  ...
)

Arguments

data
data frame with x, y, order, id, time

palette
function generating a palette (or string naming the palette)

colour
name of variable to use to specify the colour aesthetic

alpha
length two numeric, first element is the initial alpha, (optional) second is the decay rate for alpha

background
colour of the background in the plot

...
arguments to pass to geom

Value

Returns a ggplot2 object